+++ src/main/java/callprotector/spring/domain/user/service/UserServiceImpl.java
@@ -6,7 +6,7 @@ import callprotector.spring.domain.user.entity.VerificationToken;
 import callprotector.spring.domain.user.repository.UserRepository;
 import callprotector.spring.domain.user.repository.VerificationTokenRepository;
 import callprotector.spring.global.security.TokenProvider;
-import callprotector.spring.service.util.PasswordValidator;
+import callprotector.spring.global.util.PasswordValidator;
 import callprotector.spring.domain.user.dto.request.UserRequestDTO;
 import callprotector.spring.domain.user.dto.response.UserResponseDTO;
 import org.springframework.transaction.annotation.Transactional;
